make flatlist reusable and make sure that reviewer can only delete own reviews!
