+++ ReviewApp-server/src/auth/admin.guard.ts
@@ -7,7 +7,7 @@ export class AdminGuard implements CanActivate {
 
     async canActivate(context: ExecutionContext): Promise<boolean> {
         const request = context.switchToHttp().getRequest();
-        const user = request.user; // The user is already attached by JwtStrategy
+        const user = request.user; 
 
         if (!user || user.role !== 'admin') {
             throw new ForbiddenException('Access denied: You do not have permission');
